jobs_assistant/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.0
openai==1.8.0
beautifulsoup4==4.12.2
requests==2.31.0
langgraph==0.0.11
websockets==12.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
aiofiles==23.2.1 